backend/requirements.txt
fastapi
uvicorn
python-multipart
firebase-admin
python-dotenv
llama-index
llama-index-embeddings-huggingface
llama-index-llms-openai
llama-index-llms-ollama
sentence-transformers
rank_bm25
pymupdf
openai-whisper
numpy
requests
faiss-cpu
